+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+* Add extended essay to collection objects
+
+* Correct source of object essays
+
+* Correct source of artists' bios to fix missing bios
+
 ## Version 1.1
 
 * Add `collection=all` option to `Constituent#list`